tests: refactor SSL TAP tests
