+++ test/common/tls/integration/ssl_integration_test_base.h
@@ -49,12 +49,5 @@ protected:
   std::unique_ptr<ContextManager> context_manager_;
 };
 
-class SslIntegrationTest : public testing::TestWithParam<Network::Address::IpVersion>,
-                           public SslIntegrationTestBase {
-public:
-  SslIntegrationTest() : SslIntegrationTestBase(GetParam()) {}
-  void TearDown() override { SslIntegrationTestBase::TearDown(); };
-};
-
 } // namespace Ssl
 } // namespace Envoy
