+++ test/extensions/transport_sockets/tap/BUILD
@@ -22,3 +22,23 @@ envoy_extension_cc_test(
         "//test/test_common:simulated_time_system_lib",
     ],
 )
+
+envoy_extension_cc_test(
+    name = "ssl_tap_integration_test",
+    srcs = ["ssl_tap_integration_test.cc"],
+    data = [
+        "//test/config/integration/certs",
+    ],
+    extension_names = ["envoy.transport_sockets.tap"],
+    deps = [
+        "//source/common/network:connection_lib",
+        "//source/extensions/transport_sockets/tap:config",
+        "//source/extensions/transport_sockets/tls:config",
+        "//test/common/tls/integration:ssl_integration_test_lib",
+        "//test/extensions/common/tap:common",
+        "@envoy_api//envoy/config/tap/v3:pkg_cc_proto",
+        "@envoy_api//envoy/data/tap/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/transport_sockets/tap/v3:pkg_cc_proto",
+        "@envoy_api//envoy/extensions/transport_sockets/tls/v3:pkg_cc_proto",
+    ],
+)
